requirements.txt
openpyxl>=3.1.2
ttkthemes>=3.2.2
python-dotenv>=1.0.0
loguru>=0.7.0
pytest>=7.4.0
black>=23.9.0
